+++ CHANGELOG.md
@@ -6,10 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Changed 
+### Changed
 
 - Output of for LDAP user via the sub command list is now presented in a nicer ASCII table format
-- LDAP bind OU and general LDAP OU are more configurable. 
+- CLI argument for slurm, ldap and directory management can be toggled via cli and conf.toml individually
+  CLI Option for a system is only available if a sub command supports it.
+  slurm-only, ldap-only and dirs-only were replaced by new options
+- LDAP bind OU and general LDPA OU are more configurable. 
   One can now specify several dn parts like ou=people,ou=department.
   Before one could only specify a value for one fixed dn key part like people => ou=people.